django/requirements.txt
Django>=5.0,<6.0
gunicorn>=21.2
psycopg2-binary>=2.9
django-environ>=0.11
uvicorn[standard]>=0.30
djangorestframework>=3.15
django-cors-headers>=4.4
Pillow>=10.2
djangorestframework-simplejwt>=5.3
django-two-factor-auth>=1.15
django-otp>=1.3
qrcode>=7.4
django-filter==25.1.0
redis>=5.0
celery==5.4.*
django-celery-beat==2.6.*
alpaca-py